Move executable js file to bin dir. Add daemon.node & chokidar package
