+++ index.js
@@ -1,0 +1,9 @@
+const Wallet = require('./src/wallet');
+const Consumer = require('./src/consumer');
+const Producer = require('./src/producer');
+
+module.exports = {
+    Wallet,
+    Consumer,
+    Producer
+};
